+++ CHANGELOG.md
@@ -4,8 +4,8 @@
   from disk when an auto-mock is used.
 - Improve `nextjs-graphql-example`. Add a demo of an image placeholder.
 - Add Chrome Extension notifications. Config updates fire notifications.
-- Add `ping` & `shutdown` sub-commmands to `jam`
-- Add `tail-log` sub-command for tailing a current log
+- Add `ping` & `shutdown` sub-commmands to `jam-server`
+- Add `tail` sub-command for tailing a current log
 - Create a `.jambox` folder if one doesn't exist in cwd
 
 ## 0.0.9 NextJS demo, better watch mode